refactor(arc): usize sort key + expect placement in scope-exit ordering
Keys the trailing-run sort by the release var's usize index directly
and restores the too-many-arguments expectation onto the per-root gate
function it documents.

+++ compiler/ori_arc/src/aims/realize/emit_unified.rs
@@ -278,13 +278,13 @@ fn order_return_block_scope_exit_decs(func: &mut ArcFunction) {
         // One unit per release op; the sort is stable, so same-var release
         // sequences keep their relative order.
         let tail: Vec<ArcInstr> = body.split_off(start);
-        let mut units: Vec<(u32, ArcInstr)> = tail
+        let mut units: Vec<(usize, ArcInstr)> = tail
             .into_iter()
             .map(|instr| {
                 let Some(var) = release_var(&instr) else {
                     unreachable!("trailing run contains only release ops")
                 };
-                (var.index() as u32, instr)
+                (var.index(), instr)
             })
             .collect();
         units.sort_by(|a, b| b.0.cmp(&a.0));
